pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0", "poetry-dynamic-versioning",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cascoda"
version = "0.0.0"
description = "Written by the Hartree Centre Data Science group"
authors = [ "Humzah Javid <humzah.javid@stfc.ac.uk>",]
license = "MIT license"
homepage = "https://gitlab.stfc.ac.uk/HDS/cascoda"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
[[tool.mypy.overrides]]
module = [ "pandas", "numpy", "toml", "fpdf", "matplotlib.pyplot", "matplotlib",]
ignore_missing_imports = true

[tool.isort]
profile = "black"
skip = [ "build_seq.py", "conf.py",]
float_to_top = true

[tool.tox]
legacy_tox_ini = "[tox]\nskipsdist = true\nenvlist = py39\n\n[testenv]\nwhitelist_externals = poetry\npassenv = RUNNER PLATFORM_ENVVAR\ncommands =\n    poetry install -v\n    poetry run python -m unittest\n"

[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry.dependencies]
python = "~3.9"
sphinx = "^4.2.0"
sphinx-rtd-theme = "^1.0.0"
pandas = "^1.3.4"
numpy = "^1.21.4"
toml = "^0.10.2"
fpdf2 = "^2.5.1"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
isort = "^5.10.0"
jupyter = "^1.0.0"
mypy = "^0.910"
mypy-extensions = "^0.4.3"
pre-commit = "^2.15.0"
pylint = "^2.11.1"
regex = "^2021.11.2"
tox = "^3.24.4"
jupyterlab = "^3.2.9"

[tool.pylint.master]
persistent = "yes"

[tool.pylint.messages_control]
max-line-length = 88
disable = [ "wrong-hanging-indentation", "bad-whitespace",]
output-format = "colorized"
ignored_modules = [ "fpdf",]

[tool.pylint.basic]
good-names = [ "i", "j", "k", "e", "df", "n", "t",]

[tool.poetry.dev-dependencies.hooks]
git = "https://github.com/stfc/pre-commit-hooks"
tag = "v0.3.0"
